third_party/WebKit/LayoutTests/shadow-dom/css-cascade-outer-scope.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<script src="../fast/dom/shadow/resources/shadow-dom.js"></script>
<div id="sandbox"></div>
<script>
description('Outer scope rules wins, regardless of specificity.');

sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        attachShadow(
            {'mode': 'open'},
            createDOM('style', {},
                document.createTextNode(
                    '::slotted(#target) { color: red; }')),
            createDOM('slot', {})),
        createDOM('style', {},
            document.createTextNode(
                'span { color: green; }')),
        createDOM('span', {'id': 'target'})));

shouldBeEqualToString('getComputedStyle(target).color', 'rgb(0, 128, 0)');
</script>
