third_party/blink/web_tests/platform/generic/shadow-dom/css-cascade-outer-scope-expected.txt

Outer scope rules wins, regardless of specificity.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(target).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
